requirements.txt
streamlit==1.30.0
plotly==5.18.0
pandas>=1.3.0
kafka-python>=2.0.0
databricks-connect>=13.0.0
databricks-sdk>=0.20.0 